# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "DONE. API",
    "version" : "1.0.0",
    "description" : "Native Apple API. Email/password JWT and native Sign in with Apple. All timestamps UTC; history is server-owned. Durable activity events; APNs transport is not part of this phase."
  },
  "servers" : [ {
    "description" : "Local development; configure PUBLIC_ORIGIN for deployed links.",
    "url" : "http://127.0.0.1:8080"
  } ],
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      },
      "writerCapability" : {
        "description" : "43-character base64url private writer key; never a JWT and never included in the public link.",
        "type" : "http",
        "bearerFormat" : "Private write capability",
        "scheme" : "bearer"
      }
    },
    "schemas" : {
...
      "Day" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "anyOf" : [ {
              "format" : "date",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "sessions" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "additionalProperties" : false,
        "required" : [ "date", "sessions" ]
      },
...
      "FrontendStatistics" : {
        "type" : "object",
        "properties" : {
          "period" : {
            "anyOf" : [ {
              "enum" : [ "week", "month", "year" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "windowDays" : {
            "format" : "int32",
            "type" : "integer"
          },
          "from" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "to" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "timezone" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "completedSessions" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalDurationSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "averageDurationSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "mostUsedProgram" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProgramSnapshot"
            }, {
              "type" : "null"
            } ]
          },
          "changePercent" : {
            "anyOf" : [ {
              "format" : "double",
              "type" : "number"
            }, {
              "type" : "null"
            } ]
          },
          "usagePerDay" : {
            "anyOf" : [ {
              "items" : {
                "$ref" : "#/components/schemas/Day"
              },
              "type" : "array"
            }, {
              "type" : "null"
            } ]
          },
          "recentSessions" : {
            "anyOf" : [ {
              "items" : {
                "$ref" : "#/components/schemas/SessionView"
              },
              "type" : "array"
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false,
        "required" : [ "period", "windowDays", "from", "to", "timezone", "completedSessions", "totalDurationSeconds", "averageDurationSeconds", "mostUsedProgram", "changePercent", "usagePerDay", "recentSessions" ]
      },
...
      "Problem" : {
        "required" : [ "status", "code", "detail" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer"
          },
          "detail" : {
            "type" : "string"
          }
        }
      },
...
      "ProgramSnapshot" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "localizationKey" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "isCustom" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false,
        "description" : "Immutable value snapshot. name is literal user content or a built-in label; localizationKey is optional. Measured values are separate from the original snapshot.",
        "required" : [ "name", "minutes", "isCustom" ]
      },
...
      "SessionView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "applianceId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "programId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "programName" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "mode" : {
            "anyOf" : [ {
              "enum" : [ "countdown", "stopwatch" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "status" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "startedAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "expectedEnd" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "completedAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "collectedAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "canceledAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "elapsedSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          },
          "serverTime" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "sourceProgramId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "programSnapshot" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProgramSnapshot"
            }, {
              "type" : "null"
            } ]
          },
          "measuredProgram" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProgramSnapshot"
            }, {
              "type" : "null"
            } ]
          },
          "timingMode" : {
            "anyOf" : [ {
              "enum" : [ "countdown", "stopwatch" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "isOpen" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false,
        "required" : [ "id", "applianceId", "programId", "programName", "minutes", "mode", "status", "startedAt", "expectedEnd", "completedAt", "collectedAt", "canceledAt", "elapsedSeconds", "version", "serverTime", "sourceProgramId", "programSnapshot", "measuredProgram", "timingMode", "isOpen" ]
      },
...
      "Statistics" : {
        "type" : "object",
        "properties" : {
          "period" : {
            "anyOf" : [ {
              "enum" : [ "week", "month", "year" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "from" : {
            "anyOf" : [ {
              "format" : "date",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "to" : {
            "anyOf" : [ {
              "format" : "date",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "timezone" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "completedSessions" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalDurationSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "averageDurationSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "mostUsedProgram" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "changePercent" : {
            "anyOf" : [ {
              "format" : "double",
              "type" : "number"
            }, {
              "type" : "null"
            } ]
          },
          "usagePerDay" : {
            "anyOf" : [ {
              "items" : {
                "$ref" : "#/components/schemas/Day"
              },
              "type" : "array"
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false,
        "required" : [ "period", "from", "to", "timezone", "completedSessions", "totalDurationSeconds", "averageDurationSeconds", "mostUsedProgram", "changePercent", "usagePerDay" ]
      },
...
    }
  },
  "paths" : {
...
    "/api/v1/statistics" : {
      "get" : {
        "operationId" : "StatisticsController_get_1668052403",
        "tags" : [ "Statistics" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "period",
          "schema" : {
            "enum" : [ "week", "month", "year" ],
            "type" : "string",
            "default" : "week"
          },
          "required" : false,
          "in" : "query"
        }, {
          "name" : "date",
          "schema" : {
            "format" : "date",
            "type" : "string"
          },
          "required" : false,
          "in" : "query"
        }, {
          "name" : "householdId",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : false,
          "in" : "query"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Statistics"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/statistics/rolling" : {
      "get" : {
        "operationId" : "StatisticsController_rolling_2389522921",
        "tags" : [ "Statistics" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "period",
          "schema" : {
            "enum" : [ "week", "month", "year" ],
            "type" : "string",
            "default" : "week"
          },
          "required" : false,
          "in" : "query"
        }, {
          "name" : "householdId",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : false,
          "in" : "query"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FrontendStatistics"
                }
              }
            }
          }
        }
      }
    },
...
  }
}